fix: Change autocomplete to macro rather than popup
This was a personal choice because the autocomplete menu would
constantly popup, which was annoying. The macro for <C-Space> was
already setup, so I did not need to add a new macro. Instead, I just
disabled the autocomplete, which now means the menu won't pop up unless
someone uses the predefined macro.